+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Adapt: [Issue #7606 · siyuan-note/siyuan](https://github.com/siyuan-note/siyuan/issues/7606)
 - 为主题自定义配置文件添加回调函数 `callback` | Add callback function `callback` to the theme custom configuration file.
 - 调整文档树图标宽度 | Adjust the width of document tree icon.
+- 新增设置 Electron 原生主题模式功能 | Add the function of setting Electron native theme mode.
 
 ## v1.7.3/2023-03-10